+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: `exex_pragma_dispatch` implementation
 - feat: Madara ExExs proof of concept
+- fix: rejected transaction block production panic
 - fix(sync): pending block retrying mechanism
 - fix:(tests): Add testing feature to mc-db dev dependency (#294)
 - feat: new crate gateway client & server